Move DBuffer placements to placement module
