+++ megatron/core/distributed/fsdp/src/megatron_fsdp/experimental/dbuffer.py
@@ -24,29 +24,10 @@ import torch.distributed.tensor as dist_tensor
 from torch.distributed import DeviceMesh
 from torch.distributed.tensor import DTensor
 
+from .placement import Flat, Partial, Placement, Replicate, validate_placements
+
 MeshAxis = int | str
 Shape = torch.Size | Iterable[int]
-
-
-class Placement:
-    """Base class for DBuffer placements."""
-
-
-@dataclasses.dataclass(frozen=True)
-class Replicate(Placement):
-    """Replicated local buffer placement."""
-
-
-@dataclasses.dataclass(frozen=True)
-class Partial(Placement):
-    """Unreduced replicated local buffer placement."""
-
-    reduce_op: dist.ReduceOp.RedOpType = dist.ReduceOp.SUM
-
-
-@dataclasses.dataclass(frozen=True)
-class Flat(Placement):
-    """Flat per-unit dim-0 sharded local buffer placement."""
 
 
 @dataclasses.dataclass(frozen=True)
@@ -95,24 +76,6 @@ def _axis_index(mesh: DeviceMesh, axis: MeshAxis) -> int:
     if dim_names is None or axis not in dim_names:
         raise ValueError(f"Mesh axis {axis!r} is not present in mesh dim names {dim_names}.")
     return dim_names.index(axis)
-
-
-def _validate_placement(placement: Placement) -> None:
-    if not isinstance(placement, (Replicate, Partial, Flat)):
-        raise TypeError(f"Unsupported DBuffer placement: {placement!r}.")
-
-
-def _validate_placements(placements: Iterable[Placement]) -> None:
-    seen_flat = False
-    for placement in placements:
-        _validate_placement(placement)
-        if isinstance(placement, Flat):
-            seen_flat = True
-        elif seen_flat:
-            raise ValueError(
-                "Flat placements must be a suffix of the placement list so each "
-                "local buffer is a contiguous global-buffer range."
-            )
 
 
 def _compute_layout(shapes: Iterable[Shape], dp_size: int) -> GlobalLayout:
@@ -254,7 +217,7 @@ class DBuffer:
             raise ValueError(
                 f"Expected {mesh.ndim} placements for device mesh, got {len(placements)}."
             )
-        _validate_placements(placements)
+        validate_placements(placements)
 
         self.mesh = mesh
         self.placements = placements
@@ -294,7 +257,7 @@ class DBuffer:
             raise ValueError(
                 f"Expected {mesh.ndim} placements for device mesh, got {len(placements)}."
             )
-        _validate_placements(placements)
+        validate_placements(placements)
         if local_buffer.dim() != 1:
             raise ValueError("local_buffer must be a flat 1D tensor.")
 
@@ -415,7 +378,7 @@ class DBuffer:
                 f"Expected {self.mesh.ndim} placements for device mesh, got "
                 f"{len(new_placements)}."
             )
-        _validate_placements(new_placements)
+        validate_placements(new_placements)
 
         changed_axis: int | None = None
         for axis, (old_placement, new_placement) in enumerate(
@@ -461,7 +424,7 @@ class DBuffer:
 
         placements = list(self.placements)
         placements[axis] = Replicate()
-        _validate_placements(placements)
+        validate_placements(placements)
         destination = self._create_or_validate_out(placements, out)
         dist.all_gather_into_tensor(
             output_tensor=destination.local_buffer,
@@ -501,7 +464,7 @@ class DBuffer:
 
         placements = list(self.placements)
         placements[axis] = new_placement
-        _validate_placements(placements)
+        validate_placements(placements)
         destination = self._create_or_validate_out(placements, out)
         dist.reduce_scatter_tensor(
             output=destination.local_buffer,
@@ -523,7 +486,7 @@ class DBuffer:
 
         placements = list(self.placements)
         placements[axis] = new_placement
-        _validate_placements(placements)
+        validate_placements(placements)
 
         if out is None:
             destination_offset, destination_numel = self.layout.get_local_range(
